# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AcceptedBid {
  id: number;
  price: number;
  deliveryTime: string;
  coverNote: string;
  builderEmail: string;
}

export interface Bid {
  id: number;
  jobId: number;
  price: number;
  deliveryTime: string;
  coverNote: string;
  builderEmail: string;
  status: string;
  createdAt: string;
}

export interface Job {
  id: number;
  title: string;
  description: string;
  budget: number;
  deadline: string;
  category: string;
  skills: string[];
  urgent: boolean;
  isNew: boolean;
  bids: number;
  featured: boolean;
  status: string;
  clientEmail: string;
  acceptedBid?: AcceptedBid | null;
  deliveryNote?: string | null;
  deliveryLink?: string | null;
  revisionNote?: string | null;
  disputeReason?: string | null;
  stripeSessionId?: string | null;
  cancelledAt?: string | null;
  createdAt: string;
  bidsList?: Bid[];
}

export type DashboardClientJob = Job & ({
  acceptedBuilderName: string | null;
});

export type DashboardBuilderJob = Job & {
  myBid: Bid;
};

export interface CreateJobResponse {
  job: Job;
  checkoutUrl?: string | null;
}

export interface FundJobResponse {
  checkoutUrl: string;
}

export type CreateJobBodyCategory = typeof CreateJobBodyCategory[keyof typeof CreateJobBodyCategory];


export const CreateJobBodyCategory = {
  web: 'web',
  automation: 'automation',
  agent: 'agent',
  data: 'data',
} as const;

export interface CreateJobBody {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  description: string;
  /** @minimum 10 */
  budget: number;
  deadline: string;
  category: CreateJobBodyCategory;
  skills: string[];
  urgent?: boolean;
  isNew?: boolean;
  clientEmail: string;
}

export interface CreateBidBody {
  /** @minimum 1 */
  price: number;
  deliveryTime: string;
  /** @minLength 10 */
  coverNote: string;
  builderEmail: string;
}

export interface DeliverJobBody {
  builderEmail: string;
  /** @minLength 1 */
  deliveryNote: string;
  /** @minLength 1 */
  deliveryLink: string;
}

export interface ReleaseResponse {
  success: boolean;
  builderPaid: number;
  platformFee: number;
}

export interface CancelJobResponse {
  success: boolean;
  refunded: boolean;
  refundAmount?: number;
  message?: string;
}

export interface RevisionBody {
  /** @minLength 1 */
  revisionNote: string;
}

export interface DisputeBody {
  /** @minLength 1 */
  reason: string;
}

export interface Tool {
  id: number;
  name: string;
  emoji: string;
  description: string;
  price: string;
  rating: string;
  reviews: number;
  bgColor: string;
  category: string;
  priceAmount?: number | null;
  billingType?: string | null;
  builderEmail?: string | null;
  sales: number;
  status: string;
  toolUrl?: string | null;
  accessInstructions?: string | null;
  targetAudience?: string | null;
  rejectionReason?: string | null;
  createdAt: string;
}

export type CreateToolBodyBillingType = typeof CreateToolBodyBillingType[keyof typeof CreateToolBodyBillingType];


export const CreateToolBodyBillingType = {
  one_off: 'one_off',
  monthly: 'monthly',
} as const;

export interface CreateToolBody {
  name: string;
  description: string;
  /** Price in cents (e.g. 999 = $9.99) */
  price: number;
  billingType: CreateToolBodyBillingType;
  builderEmail: string;
  emoji: string;
  toolUrl?: string;
  accessInstructions: string;
  targetAudience: string;
}

export interface ToolBuyResponse {
  checkoutUrl: string;
}

export interface Builder {
  id: number;
  email: string;
  name: string;
  verified: boolean;
  stripeCustomerId?: string | null;
  stripeSubscriptionId?: string | null;
  createdAt: string;
}

export interface BuilderSubscribeBody {
  email: string;
  name?: string;
}

export interface BuilderSubscribeResponse {
  checkoutUrl: string;
}

export interface FeatureCheckoutResponse {
  checkoutUrl: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListJobsParams = {
category?: ListJobsCategory;
};

export type ListJobsCategory = typeof ListJobsCategory[keyof typeof ListJobsCategory];


export const ListJobsCategory = {
  all: 'all',
  web: 'web',
  automation: 'automation',
  agent: 'agent',
  data: 'data',
} as const;

export type DashboardClientParams = {
email: string;
};

export type DashboardBuilderParams = {
email: string;
};