# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [all, web, automation, agent, data]
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Created job with Stripe checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateJobResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/fund:
    post:
      operationId: fundJob
      tags: [jobs]
      summary: Create a Stripe Checkout session to fund a job in escrow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stripe Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FundJobResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/feature:
    post:
      operationId: featureJob
      tags: [jobs]
      summary: Create a Stripe Checkout session to feature a job listing for $15
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stripe Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureCheckoutResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/bids:
    get:
      operationId: listJobBids
      tags: [bids]
      summary: List bids for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of bids
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bid"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: submitBid
      tags: [bids]
      summary: Submit a bid on a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBidBody"
      responses:
        "201":
          description: Bid submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/bids/{bidId}/accept:
    post:
      operationId: acceptBid
      tags: [bids]
      summary: Accept a bid on a job, rejecting all others
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: bidId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job or bid not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /jobs/{id}/release:
    post:
      operationId: releaseJobPayment
      tags: [jobs]
      summary: Approve delivered work and release payment to the builder via Stripe Transfer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment released
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/revision:
    post:
      operationId: requestRevision
      tags: [jobs]
      summary: Request a revision on delivered work
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevisionBody"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/dispute:
    post:
      operationId: disputeJob
      tags: [jobs]
      summary: Open a dispute on a delivered job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeBody"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}/cancel:
    post:
      operationId: cancelJob
      tags: [jobs]
      summary: Cancel a job and refund the client before a builder is accepted
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job cancelled and refunded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelJobResponse"
        "400":
          description: Job cannot be cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    AcceptedBid:
      type: object
      properties:
        id:
          type: integer
        price:
          type: integer
        deliveryTime:
          type: string
        coverNote:
          type: string
        builderEmail:
          type: string
      required:
        - id
        - price
        - deliveryTime
        - coverNote
        - builderEmail
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        budget:
          type: integer
        deadline:
          type: string
        category:
          type: string
        skills:
          type: array
          items:
            type: string
        urgent:
          type: boolean
        isNew:
          type: boolean
        bids:
          type: integer
        featured:
          type: boolean
        status:
          type: string
        clientEmail:
          type: string
        acceptedBid:
          $ref: "#/components/schemas/AcceptedBid"
          nullable: true
        deliveryNote:
          type: string
          nullable: true
        deliveryLink:
          type: string
          nullable: true
        revisionNote:
          type: string
          nullable: true
        disputeReason:
          type: string
          nullable: true
        stripeSessionId:
          type: string
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        bidsList:
          type: array
          items:
            $ref: "#/components/schemas/Bid"
      required:
        - id
        - title
        - description
        - budget
        - deadline
        - category
        - skills
        - urgent
        - isNew
        - bids
        - featured
        - status
        - clientEmail
        - createdAt
...
    CreateJobResponse:
      type: object
      properties:
        job:
          $ref: "#/components/schemas/Job"
        checkoutUrl:
          type: string
          nullable: true
      required:
        - job
    FundJobResponse:
      type: object
      properties:
        checkoutUrl:
          type: string
      required:
        - checkoutUrl
    CreateJobBody:
      type: object
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
          minLength: 10
        budget:
          type: integer
          minimum: 10
        deadline:
          type: string
        category:
          type: string
          enum: [web, automation, agent, data]
        skills:
          type: array
          items:
            type: string
        urgent:
          type: boolean
        isNew:
          type: boolean
        clientEmail:
          type: string
          format: email
      required:
        - title
        - description
        - budget
        - deadline
        - category
        - skills
        - clientEmail
    Bid:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        price:
          type: integer
        deliveryTime:
          type: string
        coverNote:
          type: string
        builderEmail:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - jobId
        - price
        - deliveryTime
        - coverNote
        - builderEmail
        - status
        - createdAt
    CreateBidBody:
      type: object
      properties:
        price:
          type: integer
          minimum: 1
        deliveryTime:
          type: string
        coverNote:
          type: string
          minLength: 10
        builderEmail:
          type: string
          format: email
      required:
        - price
        - deliveryTime
        - coverNote
        - builderEmail
...
    ReleaseResponse:
      type: object
      properties:
        success:
          type: boolean
        builderPaid:
          type: number
        platformFee:
          type: number
      required:
        - success
        - builderPaid
        - platformFee
    CancelJobResponse:
      type: object
      properties:
        success:
          type: boolean
        refunded:
          type: boolean
        refundAmount:
          type: integer
        message:
          type: string
      required:
        - success
        - refunded
    RevisionBody:
      type: object
      properties:
        revisionNote:
          type: string
          minLength: 1
      required:
        - revisionNote
    DisputeBody:
      type: object
      properties:
        reason:
          type: string
          minLength: 1
      required:
        - reason
...
    FeatureCheckoutResponse:
      type: object
      properties:
        checkoutUrl:
          type: string
      required:
        - checkoutUrl
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error